+++ modules/hdf5_handler/HDFEOS5CFSpecialCVArray.cc
@@ -62,10 +62,10 @@ void HDFEOS5CFSpecialCVArray::read_data_NOT_from_mem_cache(bool /*add_cache*/, v
 
     bool check_pass_fileid_key = HDF5RequestHandler::get_pass_fileid();
 
-    vector<int> offset;
-    vector<int> count;
-    vector<int>step;
-    int nelms = 0;
+    vector<int64_t> offset;
+    vector<int64_t> count;
+    vector<int64_t>step;
+    int64_t nelms = 0;
 
     if (rank <= 0) 
         throw InternalErr (__FILE__, __LINE__,
